+++ 2-advanced/dubbo-samples-rpccontext/dubbo-samples-rpccontext-consumer/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.rpc.PenetrateAttachmentSelector
