+++ 2-advanced/dubbo-samples-rpccontext/dubbo-samples-rpccontext-consumer/src/main/resources/log4j2.xml
