+++ 2-advanced/dubbo-samples-rpccontext/dubbo-samples-rpccontext-consumer/src/test/resources/application.yml
@@ -1,0 +1,17 @@
+#<!-- consumer's application name, used for tracing dependency relationship (not a matching criterion),
+#don't set it same as provider -->
+#<dubbo:application name="rpccontext-consumer"/>
+#
+#<!-- use multicast registry center to discover service -->
+#<dubbo:registry address="zookeeper://${zookeeper.address:127.0.0.1}:2181"/>
+#
+#<!-- generate proxy for the remote service, then demoService can be used in the same way as the
+#local regular interface -->
+#<dubbo:reference id="consumerService" check="false" interface="org.apache.dubbo.samples.rpccontext.api.RpcContextService1" timeout="30000"/>
+dubbo:
+    application:
+        name: rpccontext-consumer
+    registry:
+        address: zookeeper://${zookeeper.address:127.0.0.1}:2181
+    protocol:
+        name: dubbo
