+++ 2-advanced/dubbo-samples-rpccontext/dubbo-samples-rpccontext-interface/pom.xml
@@ -1,0 +1,11 @@
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>org.apache.dubbo</groupId>
+        <artifactId>dubbo-samples-rpccontext</artifactId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
+    <artifactId>dubbo-samples-rpccontext-interface</artifactId>
+    <url>http://maven.apache.org</url>
+</project>
