+++ 2-advanced/dubbo-samples-rpccontext/dubbo-samples-rpccontext-provider2/src/main/java/org/apache/dubbo/samples/rpccontext/RpcContextProvider2.java
@@ -1,0 +1,38 @@
+package org.apache.dubbo.samples.rpccontext;
+/*
+ *
+ *   Licensed to the Apache Software Foundation (ASF) under one or more
+ *   contributor license agreements.  See the NOTICE file distributed with
+ *   this work for additional information regarding copyright ownership.
+ *   The ASF licenses this file to You under the Apache License, Version 2.0
+ *   (the "License"); you may not use this file except in compliance with
+ *   the License.  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+
+import java.util.logging.Logger;
+import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+@EnableDubbo
+public class RpcContextProvider2 {
+    public static final Logger LOGGER = Logger.getLogger(RpcContextProvider2.class.getName());
+
+    public static void main(String[] args) throws Exception {
+        // if you want a simple zookeeper
+        //  new EmbeddedZooKeeper(2181, false).start();
+        SpringApplication.run(RpcContextProvider2.class);
+        LOGGER.info("Rpc context provider2 started");
+    }
+
+}
