+++ 2-advanced/dubbo-samples-rpccontext/dubbo-samples-rpccontext-provider2/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.rpc.PenetrateAttachmentSelector
@@ -1,0 +1,1 @@
+mock=org.apache.dubbo.samples.rpccontext.impl.PenetrateAttachmentSelectorImpl1
