+++ 2-advanced/dubbo-samples-rpccontext/dubbo-samples-rpccontext-provider2/src/main/resources/application.yml
@@ -15,11 +15,14 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 #
-
-
-# Supported component versions of the test case
-
-# Spring app
-dubbo.version=3.3.*
-spring.version=6.*
-java.version= [ >= 17]
+dubbo:
+    application:
+        name: rpccontext-provier-2
+        qos-enable: false
+    registry:
+        address: zookeeper://${zookeeper.address:127.0.0.1}:2181
+    protocol:
+        name: dubbo
+        port: 20882
+    provider:
+        token: true
